--- a/tox.ini
+++ b/tox.ini
@@ -4,6 +4,7 @@ envlist = py35, py36, py37, {py35,py36,py37}-no-ext, lint, check
 [testenv]
 usedevelop = True
 setenv =
+    {py35,py36,py37}-no-ext: SANIC_NO_ORJSON=1
     {py35,py36,py37}-no-ext: SANIC_NO_UJSON=1
     {py35,py36,py37}-no-ext: SANIC_NO_UVLOOP=1
 deps =