Address the default JSON library replacement item mentioned in #1479 

Since https://github.com/ijl/orjson has started supporting wheel for all platforms, I suppose we can not migrate to orjson as the default library for python3.6 and above. However, for 3.5, we will still retain `ujson` until `orjson` is backported to it or we decide to drop `ujson` completely and rely on default json for python3.5