@ahopkins 2019-03-05
@harshanarayana Why not just continue to allow the developer to pick the module? Are we comfortable that `orjson` is stable enough?