@sjsadowski 2019-03-05
@ahopkins I actually started replacing ujson with orjson personally a bit ago. I'm pretty happy with it and I'm using it in production. Of course that's anecdotal, but I'd give it a pass if I was doing the review and there were no other concerns.

Personally I think sanic was already opinionated - @channelcat could have used the standard json module but went with ujson instead. Fast and correct is the way to go, and given that ujson has largely been abandoned, I think we should adopt this as a replacement and default - but still allow people to use their own json library if that's what they want to do.

This leads to a larger discussion though - do we start down the path of deprecating python 3.5 support. I'll pose that question in the community.