@sjsadowski 2019-03-06
After thinking about this some, I have two really big issues: 1) removing ujson significantly slows responses and 2) I don't want for sanic to be reliant on abandoned projects.

orjson is a young project, but it does serve a need, as illustrated below:

All of the following tests were using sanic with 12 workers (since that's how many cores I have), and running wrk with 100 concurrent connections and 2 threads.

**With ujson:**
Running 10s test @ http://localhost:5000/
  2 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     0.86ms  831.17us   7.34ms   83.63%
    Req/Sec    71.40k     7.82k   81.65k    55.50%
  1420371 requests in 10.00s, 169.32MB read
Requests/sec: 142009.29
Transfer/sec:     16.93MB

**without ujson - default library:**
Running 10s test @ http://localhost:5000/
  2 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     0.99ms    0.98ms   5.18ms   84.19%
    Req/Sec    62.65k    12.16k   79.92k    54.00%
  1246242 requests in 10.00s, 148.56MB read
Requests/sec: 124605.16
Transfer/sec:     14.85MB

**with orjson:**
Running 10s test @ http://localhost:5000/
  2 threads and 100 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     0.97ms    1.02ms   5.92ms   84.81%
    Req/Sec    69.23k     1.74k   72.03k    52.00%
  1377608 requests in 10.01s, 164.22MB read
Requests/sec: 137685.65
Transfer/sec:     16.41MB


This are just median examples, but on average over multiple runs I would see anywhere between 8-15% response time decrease (of json over ujson), and that's a poor trade-off. When going back to the standard json library, even with documentation, too many people will just ignore the docs and (might) fuss about performance. With orjson we shave that to 3-5%, and I'm sure we can hope for better optimization going forward.

So orjson gets my vote as the new default. It may be a  young project, but it's being kept up which is more than I can say about ujson.