@ijl 2019-03-06
It's important for performance that json_dumps() not convert bytes to str. There is the cost of converting bytes to str and as well the cost of converting str back to bytes when writing the request. I don't see `sanic.response.json_dumps` in the documentation. Is it meant to be private? If so, I think it can just be changed to return bytes.