@sjsadowski 2019-03-06
Thanks for the info @ijl - yes the json_dumps is a private method that can be overridden by supplying the dumps=<whatever does dumps>.

@harshanarayana can you make that update on your branch? I don't think it's significantly impacting to remove the type conversion. Someone else from @huge-success/sanic-core-devs check me to make sure I'm not wrong.