@harshanarayana 2019-03-06
@sjsadowski Sure. Let me make the changes required to use the data in bytes instead of string format. I will update this PR with the changes shortly.

@huge-success/sanic-core-devs  Can I get a No/no-go on the following items. 

1. `orjson` will be the default package used for json related ops
2. `ujson` will be retained as default on 3.5 
3. `json` is the ultimate fallback for all 3 environments