@ahopkins 2019-03-06
Is something broken in `ujson` right now? I understand the project has not had any changes, but does it need any?

If we are looking to make a change, have we looked at others? (https://github.com/python-rapidjson/python-rapidjson) Before making a change, I think we should do some more due dilligence and testing. Especially since it seems that `orjson` would **decerease** performance.