@ashleysommer 2019-03-06
I agree with @ahopkins.

I don't actually see anything wrong with sticking with ujson for now. It is fast and it works.

We will be dropping support for python 3.5 when it hits EOL in 2020-09, it would make sense to migrate to a different library then.

And additionally, as @ahopkins mentioned, why is `orjson` the intended library of choice, when high performance alternatives like `rapid-json` are available and compatible?