@yunstanford 2019-03-06
I don't have a strong opinion here, It could be included as `extra_requires`,  so `pip install sanic[orjson]`.

Also, i'd like to refactor the code a little bit. Basically, we have `compat.py`, and handle those module imports. so that in other modules, we just import from `compat.py`.