@sjsadowski 2019-03-06
So there are multiple open issues since 2017 related to number handling and deserialization issues. People have submitted PRs to address them which have been ignored.

Personally, I don't care if it's orjson or rapidjson or something else - so long as the library owner(s) respond to issues. I'd rather be ahead of problems by a voluntary change than working backward because something we rely on breaks - though to be fair we could always just drop ujson support in favor of default json. Though like I said, there's a performance hit in doing so and I'd rather replace the dead library with something actively maintained.... unless we want to fork it and maintain it ourselves, which I recommend against.