@harshanarayana 2019-03-07
@ahopkins @ashleysommer @yunstanford @sjsadowski 

I will put this PR on hold for now. Let me create a quick benchmark of a few different JSON libraries available out there and see which of them perform consistently better without having any breaking changes if possible. 

@yunstanford  I am totally for the `compat.py`. It can help us do a few different things down the line.