@ahopkins 2019-03-07
I second the `compat.py` idea from @yunstanford