@ijl 2019-03-12
orjson 2.0.2 supports python3.5.