@ahopkins 2019-04-28
@ijl In light of the [renewed discussion on bumping the min version to 3.6](https://community.sanicframework.org/t/should-we-bump-the-minimum-python-required-version-to-3-6/238/6?u=ahopkins), can you put together some new benchmarks for us to compare `ujson` with `orjson`?