@ijl 2019-04-28
There are benchmarks at https://github.com/ijl/orjson#performance.